Refactor slides repository to be based on next.js
The slides.ooni.io repository was starting to get very unmaintainable as
the node_modules needed to be duplicated (as well as every component)
for every single presentation we needed to do.

Through this refactor we are laying out every presentation inside of the
pages layout to avoid duplicating the node_modules for every
presentation and hopefully speed up the build process.

This will also make it easier to share components between the various
presentations.